+++ packages/bluesky-embed-rsc/CHANGELOG.md
@@ -1,5 +1,14 @@
 ### Unreleased:
 
+### [0.1.0] - November 26th, 2024
+
+- Fix: Fix missing line breaks within content
+- Feat: Add support for hashtag facets
+- Fix: Rendering of images if only one present
+- Fix: Click handling for embedded links
+- Fix: Rendering of link facet if truncated
+
+
 ### [0.0.1] - November 25th, 2024
 
 - Initial library release!